Move system time zone off of the clock (#99)
* Move system time zone off of the clock

* Fix missing Darwin import
