+++ core/src/commonMain/kotlin/io/islandtime/internal/Platform.kt
@@ -1,0 +1,5 @@
+package io.islandtime.internal
+
+import io.islandtime.TimeZone
+
+internal expect fun systemDefaultTimeZone(): TimeZone
